Sangaboard v0.3 support
Sangaboard v0.3 controller uses different pins to control the motors, this adds
a define to select the version.

The microcontroller on this board does not reset on new serial connections, so
a version command was addded and openflexure_stage will query the firmware
version. This means the new openflexure_stage is incompatible with older firmware.